+++ test/Transaction/CreditCardUTest.php
@@ -30,35 +30,20 @@
  * Please do not use the plugin if you do not agree to these terms of use!
  */
 
-namespace WirecardTest\PaymentSdk\Entity\PaymentMethod;
+namespace WirecardTest\PaymentSdk\Transaction;
 
-use Wirecard\PaymentSdk\Entity\PaymentMethod\ThreeDCreditCardTransaction;
+use Wirecard\PaymentSdk\Transaction\CreditCardTransaction;
+use Wirecard\PaymentSdk\Entity\Money;
 
-class ThreeDCreditCardUTest extends \PHPUnit_Framework_TestCase
+class CreditCardUTest extends \PHPUnit_Framework_TestCase
 {
-    const NOTIFICATION_TEST_URL = 'test URL';
-    const TERM_TEST_URL = 'term test URL';
-    /**
-     * @var ThreeDCreditCardTransaction
-     */
-    private $transaction;
+    const SAMPLE_TOKEN_ID = '542';
 
-    const SAMPLE_TRANSACTION_ID = '542';
-
-    public function setUp()
+    public function testGetTokenId()
     {
-        $this->transaction = new ThreeDCreditCardTransaction();
-    }
+        $ccTransaction = new CreditCardTransaction();
+        $ccTransaction->setTokenId(self::SAMPLE_TOKEN_ID);
 
-    public function testGetNotificationUrl()
-    {
-        $this->transaction->setNotificationUrl(self::NOTIFICATION_TEST_URL);
-        $this->assertEquals(self::NOTIFICATION_TEST_URL, $this->transaction->getNotificationUrl());
-    }
-
-    public function testGetTermUrl()
-    {
-        $this->transaction->setTermUrl(self::TERM_TEST_URL);
-        $this->assertEquals(self::TERM_TEST_URL, $this->transaction->getTermUrl());
+        $this->assertEquals(self::SAMPLE_TOKEN_ID, $ccTransaction->getTokenId());
     }
 }
